feat(select): done migration to JSX based options
