+++ packages/elements/dev/App.tsx
@@ -1,29 +1,64 @@
+import { For } from "solid-js";
+
 import { I18nProvider, Select } from "../src";
+
+const data = [
+  "Apple",
+  "Bacon",
+  "Banana",
+  "Broccoli",
+  "Burger",
+  "Cake",
+  "Candy",
+  "Carrot",
+  "Cherry",
+  "Chocolate",
+  "Cookie",
+  "Cucumber",
+  "Donut",
+  "Fish",
+  "Fries",
+  "Grape",
+  "Green apple",
+  "Hot dog",
+  "Ice cream",
+  "Kiwi",
+  "Lemon",
+  "Lollipop",
+  "Onion",
+  "Orange",
+  "Pasta",
+  "Pineapple",
+  "Pizza",
+  "Potato",
+  "Salad",
+  "Sandwich",
+  "Steak",
+  "Strawberry",
+  "Tomato",
+  "Watermelon",
+];
 
 function SelectDemo() {
   return (
     <Select>
-      <Select.Trigger class="select">
+      <Select.Trigger class="trigger">
         <Select.Value placeholder="Select a fruit" />
         <Select.Icon />
       </Select.Trigger>
       <Select.Portal>
         <Select.Positioner>
-          <Select.Panel class="popover">
+          <Select.Panel class="panel">
             <Select.Arrow />
-            <Select.Listbox>
-              <Select.Option value="apple" class="select-item">
-                Apple
-              </Select.Option>
-              <Select.Option value="banana" class="select-item">
-                Banana
-              </Select.Option>
-              <Select.Option value="grape" class="select-item" isDisabled>
-                Grape
-              </Select.Option>
-              <Select.Option value="orange" class="select-item">
-                Orange
-              </Select.Option>
+            <Select.Listbox class="listbox">
+              <For each={data}>
+                {item => (
+                  <Select.Option value={item} isDisabled={item === "Grape"} class="option">
+                    <Select.OptionLabel>{item}</Select.OptionLabel>
+                    <Select.OptionIndicator>X</Select.OptionIndicator>
+                  </Select.Option>
+                )}
+              </For>
             </Select.Listbox>
           </Select.Panel>
         </Select.Positioner>
