+++ packages/elements/src/select/index.ts
@@ -1,2 +1,1 @@
-export * from "./multi-select";
 export * from "./select";
