+++ packages/elements/src/select/select.test.tsx
@@ -7,18 +7,12 @@
  */
 
 import { createPointerEvent, installPointerEvent, triggerPress } from "@kobalte/tests";
+import { For } from "solid-js";
 import { fireEvent, render, screen, within } from "solid-testing-library";
 
-import { SelectBase } from "./select-base";
+import { Select } from "./select";
 
-const DATA = [
-  { label: "One", value: "one" },
-  { label: "Two", value: "two" },
-  { label: "Three", value: "three" },
-  { label: "Four", value: "four" },
-];
-
-describe("SelectBase", () => {
+describe("Select", () => {
   installPointerEvent();
 
   const onValueChange = jest.fn();
@@ -34,20 +28,23 @@ describe("SelectBase", () => {
 
   it("renders correctly", () => {
     render(() => (
-      <SelectBase options={DATA}>
-        <SelectBase.Label>Label</SelectBase.Label>
-        <SelectBase.Trigger>
-          <SelectBase.Value placeholder="Placeholder" />
-          <SelectBase.Icon />
-        </SelectBase.Trigger>
-        <SelectBase.Portal>
-          <SelectBase.Positioner>
-            <SelectBase.Menu>
-              {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-            </SelectBase.Menu>
-          </SelectBase.Positioner>
-        </SelectBase.Portal>
-      </SelectBase>
+      <Select>
+        <Select.Label>Label</Select.Label>
+        <Select.Trigger>
+          <Select.Value placeholder="Placeholder" />
+        </Select.Trigger>
+        <Select.Portal>
+          <Select.Positioner>
+            <Select.Panel>
+              <Select.Listbox>
+                <Select.Option value="1">One</Select.Option>
+                <Select.Option value="2">Two</Select.Option>
+                <Select.Option value="3">Three</Select.Option>
+              </Select.Listbox>
+            </Select.Panel>
+          </Select.Positioner>
+        </Select.Portal>
+      </Select>
     ));
 
     const select = screen.getByRole("textbox", { hidden: true });
@@ -65,102 +62,28 @@ describe("SelectBase", () => {
     expect(value).toBeVisible();
   });
 
-  it("supports custom option and group object shape", async () => {
-    const CUSTOM_DATA = [
-      {
-        key: "fruits",
-        name: "Fruits",
-        items: [
-          {
-            id: 1,
-            name: "Apple",
-            typeAheadValue: "apple",
-            isDisabled: true,
-          },
-        ],
-      },
-    ];
-
-    render(() => (
-      <SelectBase
-        options={CUSTOM_DATA}
-        optionPropertyNames={{
-          value: "id",
-          label: "name",
-          textValue: "typeAheadValue",
-          disabled: "isDisabled",
-        }}
-        optionGroupPropertyNames={{
-          id: "key",
-          label: "name",
-          options: "items",
-        }}
-      >
-        <SelectBase.Label>Label</SelectBase.Label>
-        <SelectBase.Trigger>
-          <SelectBase.Value placeholder="Placeholder" />
-          <SelectBase.Icon />
-        </SelectBase.Trigger>
-        <SelectBase.Portal>
-          <SelectBase.Positioner>
-            <SelectBase.Menu>
-              {node => (
-                <SelectBase.Group node={node()}>
-                  <SelectBase.GroupLabel data-testid="group-label">
-                    {node().label}
-                  </SelectBase.GroupLabel>
-                  <SelectBase.GroupOptions>
-                    {childNode => (
-                      <SelectBase.Option node={childNode()}>{childNode().label}</SelectBase.Option>
-                    )}
-                  </SelectBase.GroupOptions>
-                </SelectBase.Group>
-              )}
-            </SelectBase.Menu>
-          </SelectBase.Positioner>
-        </SelectBase.Portal>
-      </SelectBase>
-    ));
-
-    const trigger = screen.getByRole("button");
-    expect(trigger).toHaveTextContent("Placeholder");
-
-    await triggerPress(trigger);
-    jest.runAllTimers();
-
-    const listbox = screen.getByRole("listbox");
-
-    const groupLabel = within(listbox).getByTestId("group-label");
-
-    expect(groupLabel).toHaveTextContent("Fruits");
-
-    const items = within(listbox).getAllByRole("option");
-
-    expect(items.length).toBe(1);
-    expect(items[0]).toHaveTextContent("Apple");
-    expect(items[0]).toHaveAttribute("data-key", "1");
-    expect(items[0]).toHaveAttribute("aria-disabled", "true");
-  });
-
   describe("opening", () => {
     it("can be opened on mouse down", async () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       expect(screen.queryByRole("listbox")).toBeNull();
@@ -168,7 +91,6 @@ describe("SelectBase", () => {
       const trigger = screen.getByRole("button");
 
       await triggerPress(trigger);
-      jest.runAllTimers();
 
       const listbox = screen.getByRole("listbox");
 
@@ -180,11 +102,10 @@ describe("SelectBase", () => {
 
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[0]);
     });
@@ -193,20 +114,23 @@ describe("SelectBase", () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       expect(screen.queryByRole("listbox")).toBeNull();
@@ -241,11 +165,10 @@ describe("SelectBase", () => {
 
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[0]);
     });
@@ -254,20 +177,23 @@ describe("SelectBase", () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       expect(screen.queryByRole("listbox")).toBeNull();
@@ -290,11 +216,10 @@ describe("SelectBase", () => {
 
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[0]);
     });
@@ -303,20 +228,23 @@ describe("SelectBase", () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       expect(screen.queryByRole("listbox")).toBeNull();
@@ -339,11 +267,10 @@ describe("SelectBase", () => {
 
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[0]);
     });
@@ -352,20 +279,23 @@ describe("SelectBase", () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       expect(screen.queryByRole("listbox")).toBeNull();
@@ -388,11 +318,10 @@ describe("SelectBase", () => {
 
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[0]);
     });
@@ -401,20 +330,23 @@ describe("SelectBase", () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       expect(screen.queryByRole("listbox")).toBeNull();
@@ -437,33 +369,35 @@ describe("SelectBase", () => {
 
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
-      expect(document.activeElement).toBe(items[3]);
+      expect(document.activeElement).toBe(items[2]);
     });
 
     it("can change item focus with arrow keys", async () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       expect(screen.queryByRole("listbox")).toBeNull();
@@ -486,11 +420,10 @@ describe("SelectBase", () => {
 
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[0]);
 
@@ -517,20 +450,23 @@ describe("SelectBase", () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} isOpen onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select isOpen onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       jest.runAllTimers();
@@ -547,11 +483,10 @@ describe("SelectBase", () => {
 
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[0]);
     });
@@ -560,20 +495,23 @@ describe("SelectBase", () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} defaultIsOpen onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select defaultIsOpen onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       jest.runAllTimers();
@@ -590,11 +528,10 @@ describe("SelectBase", () => {
 
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[0]);
     });
@@ -605,20 +542,23 @@ describe("SelectBase", () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       expect(screen.queryByRole("listbox")).toBeNull();
@@ -639,7 +579,7 @@ describe("SelectBase", () => {
       await triggerPress(trigger);
       jest.runAllTimers();
 
-      expect(listbox).not.toBeInTheDocument();
+      expect(listbox).not.toBeVisible();
       expect(trigger).toHaveAttribute("aria-expanded", "false");
       expect(trigger).not.toHaveAttribute("aria-controls");
       expect(onOpenChange).toBeCalledTimes(2);
@@ -652,20 +592,23 @@ describe("SelectBase", () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       expect(screen.queryByRole("listbox")).toBeNull();
@@ -686,7 +629,7 @@ describe("SelectBase", () => {
       await triggerPress(document.body);
       jest.runAllTimers();
 
-      expect(listbox).not.toBeInTheDocument();
+      expect(listbox).not.toBeVisible();
       expect(trigger).toHaveAttribute("aria-expanded", "false");
       expect(trigger).not.toHaveAttribute("aria-controls");
       expect(onOpenChange).toBeCalledTimes(2);
@@ -697,20 +640,23 @@ describe("SelectBase", () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       expect(screen.queryByRole("listbox")).toBeNull();
@@ -730,7 +676,7 @@ describe("SelectBase", () => {
       fireEvent.keyDown(listbox, { key: "Escape" });
       await Promise.resolve();
 
-      expect(listbox).not.toBeInTheDocument();
+      expect(listbox).not.toBeVisible();
       expect(trigger).toHaveAttribute("aria-expanded", "false");
       expect(trigger).not.toHaveAttribute("aria-controls");
       expect(onOpenChange).toBeCalledTimes(2);
@@ -743,20 +689,23 @@ describe("SelectBase", () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} isOpen onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select isOpen onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       jest.runAllTimers();
@@ -785,20 +734,23 @@ describe("SelectBase", () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} defaultIsOpen onOpenChange={onOpenChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select defaultIsOpen onOpenChange={onOpenChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       jest.runAllTimers();
@@ -818,7 +770,7 @@ describe("SelectBase", () => {
       await Promise.resolve();
       jest.runAllTimers();
 
-      expect(listbox).not.toBeInTheDocument();
+      expect(listbox).not.toBeVisible();
       expect(trigger).toHaveAttribute("aria-expanded", "false");
       expect(trigger).not.toHaveAttribute("aria-controls");
       expect(onOpenChange).toBeCalledTimes(1);
@@ -829,20 +781,23 @@ describe("SelectBase", () => {
   describe("labeling", () => {
     it("focuses on the trigger when you click the label", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const label = screen.getAllByText("Label")[0];
@@ -857,20 +812,23 @@ describe("SelectBase", () => {
 
     it("supports labeling with a visible label", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -894,19 +852,22 @@ describe("SelectBase", () => {
 
     it("supports labeling via aria-label", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Trigger aria-label="foo">
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Trigger aria-label="foo">
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -928,19 +889,22 @@ describe("SelectBase", () => {
 
     it("supports labeling via aria-labelledby", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Trigger aria-labelledby="foo">
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Trigger aria-labelledby="foo">
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -961,19 +925,22 @@ describe("SelectBase", () => {
 
     it("supports labeling via aria-label and aria-labelledby", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Trigger aria-label="bar" aria-labelledby="foo">
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Trigger aria-label="bar" aria-labelledby="foo">
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -996,21 +963,24 @@ describe("SelectBase", () => {
   describe("help text", () => {
     it("supports description", () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Description>Description</SelectBase.Description>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Description>Description</Select.Description>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       jest.runAllTimers();
@@ -1024,21 +994,24 @@ describe("SelectBase", () => {
 
     it("supports error message", () => {
       render(() => (
-        <SelectBase options={DATA} validationState="invalid" onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.ErrorMessage>ErrorMessage</SelectBase.ErrorMessage>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select validationState="invalid" onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.ErrorMessage>ErrorMessage</Select.ErrorMessage>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       jest.runAllTimers();
@@ -1054,47 +1027,48 @@ describe("SelectBase", () => {
   describe("selection", () => {
     it("can select items on press", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
       expect(trigger).toHaveTextContent("Placeholder");
 
       await triggerPress(trigger);
-      jest.runAllTimers();
 
       const listbox = screen.getByRole("listbox");
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[0]);
 
       await triggerPress(items[2]);
 
       expect(onValueChange).toHaveBeenCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("three")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("3")).toBeTruthy();
 
       jest.runAllTimers();
 
-      expect(listbox).not.toBeInTheDocument();
+      expect(listbox).not.toBeVisible();
 
       // run restore focus rAF
       jest.runAllTimers();
@@ -1103,114 +1077,39 @@ describe("SelectBase", () => {
       expect(trigger).toHaveTextContent("Three");
     });
 
-    it("can select items with falsy keys", async () => {
-      const FALSY_KEY_DATA = [
-        { value: "", label: "Empty" },
-        { value: 0, label: "Zero" },
-      ];
-
-      render(() => (
-        <SelectBase options={FALSY_KEY_DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
-      ));
-
-      const trigger = screen.getByRole("button");
-
-      expect(trigger).toHaveTextContent("Placeholder");
-
-      await triggerPress(trigger);
-      jest.runAllTimers();
-
-      let listbox = screen.getByRole("listbox");
-      const items = within(listbox).getAllByRole("option");
-
-      expect(items.length).toBe(2);
-      expect(items[0]).toHaveTextContent("Empty");
-      expect(items[1]).toHaveTextContent("Zero");
-
-      expect(document.activeElement).toBe(items[0]);
-
-      await triggerPress(items[0]);
-
-      expect(onValueChange).toHaveBeenCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("")).toBeTruthy();
-
-      jest.runAllTimers();
-
-      expect(listbox).not.toBeInTheDocument();
-
-      // run restore focus rAF
-      jest.runAllTimers();
-
-      expect(document.activeElement).toBe(trigger);
-      expect(trigger).toHaveTextContent("Empty");
-
-      await triggerPress(trigger);
-      jest.runAllTimers();
-
-      listbox = screen.getByRole("listbox");
-      const item1 = within(listbox).getByText("Zero");
-
-      await triggerPress(item1);
-
-      expect(onValueChange).toHaveBeenCalledTimes(2);
-      expect(onValueChange.mock.calls[1][0].has(0)).toBeTruthy();
-
-      jest.runAllTimers();
-
-      expect(listbox).not.toBeInTheDocument();
-
-      // run restore focus rAF
-      jest.runAllTimers();
-
-      expect(document.activeElement).toBe(trigger);
-      expect(trigger).toHaveTextContent("Zero");
-    });
-
     it("can select items with the Space key", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
       expect(trigger).toHaveTextContent("Placeholder");
 
       await triggerPress(trigger);
-      jest.runAllTimers();
 
       const listbox = screen.getByRole("listbox");
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[0]);
 
@@ -1229,11 +1128,11 @@ describe("SelectBase", () => {
       await Promise.resolve();
 
       expect(onValueChange).toHaveBeenCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("two")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("2")).toBeTruthy();
 
       jest.runAllTimers();
 
-      expect(listbox).not.toBeInTheDocument();
+      expect(listbox).not.toBeVisible();
 
       // run restore focus rAF
       jest.runAllTimers();
@@ -1244,20 +1143,23 @@ describe("SelectBase", () => {
 
     it("can select items with the Enter key", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -1272,18 +1174,15 @@ describe("SelectBase", () => {
       fireEvent.keyUp(trigger, { key: "ArrowUp" });
       await Promise.resolve();
 
-      jest.runAllTimers();
-
       const listbox = screen.getByRole("listbox");
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
-      expect(document.activeElement).toBe(items[3]);
+      expect(document.activeElement).toBe(items[2]);
 
       fireEvent.keyDown(listbox, { key: "ArrowUp" });
       await Promise.resolve();
@@ -1291,7 +1190,7 @@ describe("SelectBase", () => {
       fireEvent.keyUp(listbox, { key: "ArrowUp" });
       await Promise.resolve();
 
-      expect(document.activeElement).toBe(items[2]);
+      expect(document.activeElement).toBe(items[1]);
 
       fireEvent.keyDown(document.activeElement!, { key: "Enter" });
       await Promise.resolve();
@@ -1300,46 +1199,47 @@ describe("SelectBase", () => {
       await Promise.resolve();
 
       expect(onValueChange).toHaveBeenCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("three")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("2")).toBeTruthy();
 
       jest.runAllTimers();
 
-      expect(listbox).not.toBeInTheDocument();
-      expect(trigger).toHaveTextContent("Three");
+      expect(listbox).not.toBeVisible();
+      expect(trigger).toHaveTextContent("Two");
     });
 
     it("focuses items on hover", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
       expect(trigger).toHaveTextContent("Placeholder");
 
       await triggerPress(trigger);
-      jest.runAllTimers();
 
       const listbox = screen.getByRole("listbox");
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[0]);
 
@@ -1366,11 +1266,11 @@ describe("SelectBase", () => {
       await Promise.resolve();
 
       expect(onValueChange).toHaveBeenCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("three")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("3")).toBeTruthy();
 
       jest.runAllTimers();
 
-      expect(listbox).not.toBeInTheDocument();
+      expect(listbox).not.toBeVisible();
 
       // run restore focus rAF
       jest.runAllTimers();
@@ -1382,20 +1282,23 @@ describe("SelectBase", () => {
     it("does not clear selection on escape closing the listbox", async () => {
       const onOpenChangeSpy = jest.fn();
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange} onOpenChange={onOpenChangeSpy}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange} onOpenChange={onOpenChangeSpy}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -1458,20 +1361,23 @@ describe("SelectBase", () => {
 
     it("supports controlled selection", async () => {
       render(() => (
-        <SelectBase options={DATA} value={["two"]} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select value={["2"]} onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -1483,11 +1389,10 @@ describe("SelectBase", () => {
       const listbox = screen.getByRole("listbox");
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[1]);
 
@@ -1508,11 +1413,11 @@ describe("SelectBase", () => {
       await Promise.resolve();
 
       expect(onValueChange).toHaveBeenCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("one")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("1")).toBeTruthy();
 
       jest.runAllTimers();
 
-      expect(listbox).not.toBeInTheDocument();
+      expect(listbox).not.toBeVisible();
 
       // run restore focus rAF
       jest.runAllTimers();
@@ -1523,20 +1428,23 @@ describe("SelectBase", () => {
 
     it("supports default selection", async () => {
       render(() => (
-        <SelectBase options={DATA} defaultValue={["two"]} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select defaultValue={["2"]} onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -1549,11 +1457,10 @@ describe("SelectBase", () => {
       const listbox = screen.getByRole("listbox");
       const items = within(listbox).getAllByRole("option");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[1]);
       expect(items[1]).toHaveAttribute("aria-selected", "true");
@@ -1573,11 +1480,11 @@ describe("SelectBase", () => {
       await Promise.resolve();
 
       expect(onValueChange).toHaveBeenCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("one")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("1")).toBeTruthy();
 
       jest.runAllTimers();
 
-      expect(listbox).not.toBeInTheDocument();
+      expect(listbox).not.toBeVisible();
 
       // run restore focus rAF
       jest.runAllTimers();
@@ -1587,34 +1494,32 @@ describe("SelectBase", () => {
     });
 
     it("skips disabled items", async () => {
-      const WITH_DISABLED_DATA = [
-        { label: "One", value: "one" },
-        { label: "Two", value: "two", disabled: true },
-        { label: "Three", value: "three" },
-      ];
-
       render(() => (
-        <SelectBase options={WITH_DISABLED_DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2" isDisabled>
+                    Two
+                  </Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
       expect(trigger).toHaveTextContent("Placeholder");
 
       await triggerPress(trigger);
-      jest.runAllTimers();
 
       const listbox = screen.getByRole("listbox");
       const items = within(listbox).getAllByRole("option");
@@ -1642,11 +1547,11 @@ describe("SelectBase", () => {
       await Promise.resolve();
 
       expect(onValueChange).toHaveBeenCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("three")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("3")).toBeTruthy();
 
       jest.runAllTimers();
 
-      expect(listbox).not.toBeInTheDocument();
+      expect(listbox).not.toBeVisible();
 
       // run restore focus rAF
       jest.runAllTimers();
@@ -1657,20 +1562,24 @@ describe("SelectBase", () => {
 
     it("supports type to select", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                  <Select.Option value="4">Four</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -1719,11 +1628,11 @@ describe("SelectBase", () => {
       await Promise.resolve();
 
       expect(onValueChange).toHaveBeenCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("three")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("3")).toBeTruthy();
 
       jest.runAllTimers();
 
-      expect(listbox).not.toBeInTheDocument();
+      expect(listbox).not.toBeVisible();
       expect(trigger).toHaveTextContent("Three");
 
       jest.advanceTimersByTime(500);
@@ -1751,35 +1660,37 @@ describe("SelectBase", () => {
 
       jest.runAllTimers();
 
-      expect(listbox).not.toBeInTheDocument();
+      expect(listbox).not.toBeVisible();
       expect(trigger).toHaveTextContent("Four");
       expect(onValueChange).toHaveBeenCalledTimes(2);
-      expect(onValueChange.mock.calls[1][0].has("four")).toBeTruthy();
+      expect(onValueChange.mock.calls[1][0].has("4")).toBeTruthy();
     });
 
     it("does not deselect when pressing an already selected item", async () => {
       render(() => (
-        <SelectBase options={DATA} defaultValue={["two"]} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select defaultValue={["2"]} onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
       expect(trigger).toHaveTextContent("Two");
 
       await triggerPress(trigger);
-      jest.runAllTimers();
 
       const listbox = screen.getByRole("listbox");
       const items = within(listbox).getAllByRole("option");
@@ -1789,11 +1700,11 @@ describe("SelectBase", () => {
       await triggerPress(items[1]);
 
       expect(onValueChange).toHaveBeenCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("two")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("2")).toBeTruthy();
 
       jest.runAllTimers();
 
-      expect(listbox).not.toBeInTheDocument();
+      expect(listbox).not.toBeVisible();
 
       // run restore focus rAF
       jest.runAllTimers();
@@ -1804,20 +1715,23 @@ describe("SelectBase", () => {
 
     it("move selection on Arrow-Left/Right", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -1830,8 +1744,6 @@ describe("SelectBase", () => {
       fireEvent.keyDown(trigger, { key: "ArrowLeft" });
       await Promise.resolve();
 
-      jest.runAllTimers();
-
       expect(onValueChange).toHaveBeenCalledTimes(1);
       expect(trigger).toHaveTextContent("One");
 
@@ -1841,6 +1753,7 @@ describe("SelectBase", () => {
       expect(trigger).toHaveTextContent("One");
 
       fireEvent.keyDown(trigger, { key: "ArrowRight" });
+
       expect(onValueChange).toHaveBeenCalledTimes(2);
       expect(trigger).toHaveTextContent("Two");
 
@@ -1853,19 +1766,12 @@ describe("SelectBase", () => {
       fireEvent.keyDown(trigger, { key: "ArrowRight" });
       await Promise.resolve();
 
-      expect(onValueChange).toHaveBeenCalledTimes(4);
-      expect(trigger).toHaveTextContent("Four");
-
-      fireEvent.keyDown(trigger, { key: "ArrowLeft" });
-      await Promise.resolve();
-
-      expect(onValueChange).toHaveBeenCalledTimes(5);
       expect(trigger).toHaveTextContent("Three");
 
       fireEvent.keyDown(trigger, { key: "ArrowLeft" });
       await Promise.resolve();
 
-      expect(onValueChange).toHaveBeenCalledTimes(6);
+      expect(onValueChange).toHaveBeenCalledTimes(4);
       expect(trigger).toHaveTextContent("Two");
     });
   });
@@ -1873,38 +1779,39 @@ describe("SelectBase", () => {
   describe("multi selection", () => {
     it("supports selecting multiple options", async () => {
       render(() => (
-        <SelectBase options={DATA} isMultiple onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select selectionMode="multiple" onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
       expect(trigger).toHaveTextContent("Placeholder");
 
       await triggerPress(trigger);
-      jest.runAllTimers();
 
       const listbox = screen.getByRole("listbox");
       const items = within(listbox).getAllByRole("option");
 
       expect(listbox).toHaveAttribute("aria-multiselectable", "true");
 
-      expect(items.length).toBe(4);
+      expect(items.length).toBe(3);
       expect(items[0]).toHaveTextContent("One");
       expect(items[1]).toHaveTextContent("Two");
       expect(items[2]).toHaveTextContent("Three");
-      expect(items[3]).toHaveTextContent("Four");
 
       expect(document.activeElement).toBe(items[0]);
 
@@ -1915,38 +1822,36 @@ describe("SelectBase", () => {
       expect(items[2]).toHaveAttribute("aria-selected", "true");
 
       expect(onValueChange).toBeCalledTimes(2);
-      expect(onValueChange.mock.calls[0][0].has("one")).toBeTruthy();
-      expect(onValueChange.mock.calls[1][0].has("three")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("1")).toBeTruthy();
+      expect(onValueChange.mock.calls[1][0].has("3")).toBeTruthy();
 
       // Does not close on multi-select
-      expect(listbox).toBeInTheDocument();
+      expect(listbox).toBeVisible();
 
       expect(trigger).toHaveTextContent("One, Three");
     });
 
     it("supports multiple defaultValue (uncontrolled)", async () => {
-      const defaultValue = new Set(["one", "two"]);
+      const defaultValue = new Set(["1", "2"]);
 
       render(() => (
-        <SelectBase
-          options={DATA}
-          isMultiple
-          defaultValue={defaultValue}
-          onValueChange={onValueChange}
-        >
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select selectionMode="multiple" defaultValue={defaultValue} onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -1967,29 +1872,32 @@ describe("SelectBase", () => {
       expect(items[2]).toHaveAttribute("aria-selected", "true");
 
       expect(onValueChange).toBeCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("one")).toBeTruthy();
-      expect(onValueChange.mock.calls[0][0].has("two")).toBeTruthy();
-      expect(onValueChange.mock.calls[0][0].has("three")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("1")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("2")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("3")).toBeTruthy();
     });
 
     it("supports multiple value (controlled)", async () => {
-      const value = new Set(["one", "two"]);
+      const value = new Set(["1", "2"]);
 
       render(() => (
-        <SelectBase options={DATA} isMultiple value={value} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select selectionMode="multiple" value={value} onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -2010,32 +1918,30 @@ describe("SelectBase", () => {
       expect(items[2]).toHaveAttribute("aria-selected", "false");
 
       expect(onValueChange).toBeCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("three")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("3")).toBeTruthy();
     });
 
     it("supports deselection", async () => {
-      const defaultValue = new Set(["one", "two"]);
+      const defaultValue = new Set(["1", "2"]);
 
       render(() => (
-        <SelectBase
-          options={DATA}
-          isMultiple
-          defaultValue={defaultValue}
-          onValueChange={onValueChange}
-        >
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select selectionMode="multiple" defaultValue={defaultValue} onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -2055,28 +1961,31 @@ describe("SelectBase", () => {
       expect(items[0]).toHaveAttribute("aria-selected", "false");
 
       expect(onValueChange).toBeCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("one")).toBeFalsy();
-      expect(onValueChange.mock.calls[0][0].has("two")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("1")).toBeFalsy();
+      expect(onValueChange.mock.calls[0][0].has("2")).toBeTruthy();
     });
   });
 
   describe("type to select", () => {
     it("supports focusing items by typing letters in rapid succession without opening the menu", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -2093,7 +2002,7 @@ describe("SelectBase", () => {
       await Promise.resolve();
 
       expect(onValueChange).toHaveBeenCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("two")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("2")).toBeTruthy();
       expect(trigger).toHaveTextContent("Two");
 
       fireEvent.keyDown(trigger, { key: "h" });
@@ -2103,26 +2012,29 @@ describe("SelectBase", () => {
       await Promise.resolve();
 
       expect(onValueChange).toHaveBeenCalledTimes(2);
-      expect(onValueChange.mock.calls[1][0].has("three")).toBeTruthy();
+      expect(onValueChange.mock.calls[1][0].has("3")).toBeTruthy();
       expect(trigger).toHaveTextContent("Three");
     });
 
     it("resets the search text after a timeout", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -2139,7 +2051,7 @@ describe("SelectBase", () => {
       await Promise.resolve();
 
       expect(onValueChange).toHaveBeenCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("two")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("2")).toBeTruthy();
       expect(trigger).toHaveTextContent("Two");
 
       jest.runAllTimers();
@@ -2156,20 +2068,23 @@ describe("SelectBase", () => {
 
     it("wraps around when no items past the current one match", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
@@ -2185,7 +2100,7 @@ describe("SelectBase", () => {
       await Promise.resolve();
 
       expect(onValueChange).toHaveBeenCalledTimes(1);
-      expect(onValueChange.mock.calls[0][0].has("two")).toBeTruthy();
+      expect(onValueChange.mock.calls[0][0].has("2")).toBeTruthy();
       expect(trigger).toHaveTextContent("Two");
 
       jest.runAllTimers();
@@ -2210,39 +2125,30 @@ describe("SelectBase", () => {
       ];
 
       render(() => (
-        <SelectBase
-          options={ADDRESS_DATA}
-          autoComplete="address-level1"
-          onValueChange={onValueChange}
-        >
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select autoComplete="address-level1" onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <For each={ADDRESS_DATA}>
+                    {data => <Select.Option value={data.value}>{data.label}</Select.Option>}
+                  </For>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const trigger = screen.getByRole("button");
 
       expect(trigger).toHaveTextContent("Placeholder");
 
-      // For anyone else who comes through this listbox/combobox path
-      // I can't use combobox here because there is a size attribute on the html select
-      // everything below this line is the path i followed to get to the correct role:
-      //   not sure why i can't use listbox https://github.com/A11yance/aria-query#elements-to-roles
-      //   however, i think this is correct based on https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles
-      //   which says "The listbox role is used for lists from which a user may select one or more items which are static and, unlike HTML <select> elements, may contain images."
-      //   Also, this test in react testing library seems to indicate something about size which we do not currently have, probably a bug
-      //   https://github.com/testing-library/dom-testing-library/blob/master/src/__tests__/element-queries.js#L548
-      const hiddenSelectBase = screen.getByRole("listbox", { hidden: true });
+      const hiddenSelectBase = screen.getAllByRole("listbox", { hidden: true })[0];
 
       expect(hiddenSelectBase).toHaveAttribute("tabIndex", "-1");
       expect(hiddenSelectBase).toHaveAttribute("autocomplete", "address-level1");
@@ -2267,20 +2173,23 @@ describe("SelectBase", () => {
     // TODO: failing, don't know why.
     it.skip("should have a hidden input to marshall focus to the button", async () => {
       render(() => (
-        <SelectBase options={DATA} onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const hiddenInput = screen.getByRole("textbox", { hidden: true }); // get the hidden ones
@@ -2307,20 +2216,23 @@ describe("SelectBase", () => {
   describe("disabled", () => {
     it("disables the hidden select when isDisabled is true", async () => {
       render(() => (
-        <SelectBase options={DATA} isDisabled onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select isDisabled onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       const select = screen.getByRole("textbox", { hidden: true });
@@ -2332,20 +2244,23 @@ describe("SelectBase", () => {
       const onOpenChange = jest.fn();
 
       render(() => (
-        <SelectBase options={DATA} isDisabled onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select isDisabled onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       expect(screen.queryByRole("listbox")).toBeNull();
@@ -2365,20 +2280,23 @@ describe("SelectBase", () => {
     it("does not open on Space key press when isDisabled is true", async () => {
       const onOpenChange = jest.fn();
       render(() => (
-        <SelectBase options={DATA} isDisabled onValueChange={onValueChange}>
-          <SelectBase.Label>Label</SelectBase.Label>
-          <SelectBase.Trigger>
-            <SelectBase.Value placeholder="Placeholder" />
-            <SelectBase.Icon />
-          </SelectBase.Trigger>
-          <SelectBase.Portal>
-            <SelectBase.Positioner>
-              <SelectBase.Menu>
-                {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-              </SelectBase.Menu>
-            </SelectBase.Positioner>
-          </SelectBase.Portal>
-        </SelectBase>
+        <Select isDisabled onValueChange={onValueChange}>
+          <Select.Label>Label</Select.Label>
+          <Select.Trigger>
+            <Select.Value placeholder="Placeholder" />
+          </Select.Trigger>
+          <Select.Portal>
+            <Select.Positioner>
+              <Select.Panel>
+                <Select.Listbox>
+                  <Select.Option value="1">One</Select.Option>
+                  <Select.Option value="2">Two</Select.Option>
+                  <Select.Option value="3">Three</Select.Option>
+                </Select.Listbox>
+              </Select.Panel>
+            </Select.Positioner>
+          </Select.Portal>
+        </Select>
       ));
 
       expect(screen.queryByRole("listbox")).toBeNull();
@@ -2414,20 +2332,23 @@ describe("SelectBase", () => {
 
       render(() => (
         <form data-testid="form" onSubmit={onSubmit}>
-          <SelectBase name="test" options={DATA}>
-            <SelectBase.Label>Label</SelectBase.Label>
-            <SelectBase.Trigger autofocus>
-              <SelectBase.Value placeholder="Placeholder" />
-              <SelectBase.Icon />
-            </SelectBase.Trigger>
-            <SelectBase.Portal>
-              <SelectBase.Positioner>
-                <SelectBase.Menu>
-                  {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-                </SelectBase.Menu>
-              </SelectBase.Positioner>
-            </SelectBase.Portal>
-          </SelectBase>
+          <Select name="test">
+            <Select.Label>Label</Select.Label>
+            <Select.Trigger autofocus>
+              <Select.Value placeholder="Placeholder" />
+            </Select.Trigger>
+            <Select.Portal>
+              <Select.Positioner>
+                <Select.Panel>
+                  <Select.Listbox>
+                    <Select.Option value="1">One</Select.Option>
+                    <Select.Option value="2">Two</Select.Option>
+                    <Select.Option value="3">Three</Select.Option>
+                  </Select.Listbox>
+                </Select.Panel>
+              </Select.Positioner>
+            </Select.Portal>
+          </Select>
           <button type="submit" data-testid="submit">
             submit
           </button>
@@ -2452,20 +2373,23 @@ describe("SelectBase", () => {
 
       render(() => (
         <form data-testid="form" onSubmit={onSubmit}>
-          <SelectBase name="test" defaultValue={["one"]} options={DATA}>
-            <SelectBase.Label>Label</SelectBase.Label>
-            <SelectBase.Trigger autofocus>
-              <SelectBase.Value placeholder="Placeholder" />
-              <SelectBase.Icon />
-            </SelectBase.Trigger>
-            <SelectBase.Portal>
-              <SelectBase.Positioner>
-                <SelectBase.Menu>
-                  {node => <SelectBase.Option node={node()}>{node().label}</SelectBase.Option>}
-                </SelectBase.Menu>
-              </SelectBase.Positioner>
-            </SelectBase.Portal>
-          </SelectBase>
+          <Select name="test" defaultValue={["1"]}>
+            <Select.Label>Label</Select.Label>
+            <Select.Trigger autofocus>
+              <Select.Value placeholder="Placeholder" />
+            </Select.Trigger>
+            <Select.Portal>
+              <Select.Positioner>
+                <Select.Panel>
+                  <Select.Listbox>
+                    <Select.Option value="1">One</Select.Option>
+                    <Select.Option value="2">Two</Select.Option>
+                    <Select.Option value="3">Three</Select.Option>
+                  </Select.Listbox>
+                </Select.Panel>
+              </Select.Positioner>
+            </Select.Portal>
+          </Select>
         </form>
       ));
 
@@ -2473,7 +2397,7 @@ describe("SelectBase", () => {
       await Promise.resolve();
 
       expect(onSubmit).toHaveBeenCalledTimes(1);
-      expect(value).toEqual("one");
+      expect(value).toEqual("1");
     });
   });
 });
